MHC-5,6: refactoring (kts migration is for future)
