--- baseUI/src/main/java/ru/marvelheroes/extensions/IntExt.kt
@@ -1,4 +1,0 @@
-package ru.marvelheroes.extensions
-
-const val MILLIS = 1000
-const val QUALITY = 100
